Refactor: Move SiteBuilder to new.rs file
This makes this code public and implementation code can be more private

Formating fixes
Refactor: Remove unused imports
refactor: remove unneeded code
Refactor: time::Date added to helps
Returns a time:Date object that an be set by default in the
MetaArticle Struct
Exploring how to handle adding authors
Move related code closer together
